api/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-pptx==0.6.23
Pillow==10.2.0
requests==2.31.0
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.4.2
